+++ libraries/matchers.rb
@@ -4,7 +4,7 @@
 #
 # Copyright:: Copyright (c) 2014, Voxer LLC
 # License: MIT License
-# Author: Dave Eddy <dave@daveeddy.com
+# Author: Dave Eddy <dave@daveeddy.com>
 #
 
 if defined?(ChefSpec)
