+++ recipes/default.rb
@@ -4,5 +4,5 @@
 #
 # Copyright:: Copyright (c) 2014, Voxer LLC
 # License: MIT License
-# Author: Dave Eddy <dave@daveeddy.com
+# Author: Dave Eddy <dave@daveeddy.com>
 #
